refactor(fusion): the failure helpers are their own module
A pure move out of `ordering.rs`, which had grown four unrelated
concepts. `panic_message` in particular is a panic-payload utility with
nothing to do with execution ordering.
