+++ crates/burn-fusion/src/stream/execution/mod.rs
@@ -2,6 +2,7 @@ pub(crate) mod validator;
 
 mod base;
 mod explorer;
+mod failure;
 mod ordering;
 mod policy;
 mod processor;
@@ -11,6 +12,8 @@ pub(crate) use trace::{log_execution_table, op_kind};
 
 pub use base::*;
 pub use ordering::*;
+
+pub(crate) use failure::*;
 
 pub(crate) use explorer::*;
 pub(crate) use policy::*;
